requirements.txt
# torch==2.5.1+cu121
# --extra-index-url https://download.pytorch.org/whl/cu121
# unsloth
# vllm
# huggingface_hub
# python-dotenv


--extra-index-url https://download.pytorch.org/whl/cu121
torch
unsloth
vllm==0.11.2
huggingface_hub
python-dotenv
wandb
weave




